timeutils: move timeutils.{c,h} to lib/timeutils
