+++ lib/CMakeLists.txt
@@ -124,7 +124,6 @@ set (LIB_HEADERS
     syslog-names.h
     syslog-ng.h
     string-list.h
-    timeutils.h
     tls-support.h
     thread-utils.h
     tlscontext.h
@@ -211,7 +210,6 @@ set(LIB_SOURCES
     str-utils.c
     syslog-names.c
     string-list.c
-    timeutils.c
     type-hinting.c
     ringbuffer.c
     late_ack_tracker.c
