+++ lib/apphook.c
@@ -29,7 +29,7 @@
 #include "alarms.h"
 #include "stats/stats-registry.h"
 #include "logmsg/logmsg.h"
-#include "timeutils.h"
+#include "timeutils/timeutils.h"
 #include "logsource.h"
 #include "logwriter.h"
 #include "afinter.h"
