+++ lib/logsource.c
@@ -25,7 +25,7 @@
 #include "logsource.h"
 #include "messages.h"
 #include "host-resolve.h"
-#include "timeutils.h"
+#include "timeutils/timeutils.h"
 #include "stats/stats-registry.h"
 #include "msg-stats.h"
 #include "logmsg/tags.h"
