+++ lib/stats/stats.c
@@ -27,7 +27,7 @@
 #include "stats/stats-query.h"
 #include "stats/stats-registry.h"
 #include "stats/stats.h"
-#include "timeutils.h"
+#include "timeutils/timeutils.h"
 
 #include <string.h>
 #include <iv.h>
