+++ lib/timeutils/Makefile.am
@@ -3,9 +3,11 @@ timeutilsincludedir = ${pkgincludedir}/timeutils
 EXTRA_DIST += lib/timeutils/CMakeLists.txt
 
 timeutilsinclude_HEADERS =	\
-  lib/timeutils/decode.h
+  lib/timeutils/decode.h	\
+  lib/timeutils/timeutils.h
 
 timeutils_sources =		\
- lib/timeutils/decode.c
+ lib/timeutils/decode.c		\
+ lib/timeutils/timeutils.c
 
 include lib/timeutils/tests/Makefile.am
