+++ modules/afsql/afsql.c
@@ -31,7 +31,7 @@
 #include "seqnum.h"
 #include "stats/stats-registry.h"
 #include "apphook.h"
-#include "timeutils.h"
+#include "timeutils/timeutils.h"
 #include "mainloop-worker.h"
 #include "str-utils.h"
