+++ modules/syslogformat/syslog-format.c
@@ -25,7 +25,7 @@
 #include "timeutils/decode.h"
 #include "logmsg/logmsg.h"
 #include "messages.h"
-#include "timeutils.h"
+#include "timeutils/timeutils.h"
 #include "find-crlf.h"
 #include "cfg.h"
 #include "str-format.h"
