+++ tests/unit/test_logwriter.c
@@ -27,7 +27,7 @@
 #include "template/templates.h"
 #include "apphook.h"
 #include "cfg.h"
-#include "timeutils.h"
+#include "timeutils/timeutils.h"
 #include "plugin.h"
 #include "logqueue-fifo.h"
