+++ tests/unit/test_zone.c
@@ -22,8 +22,8 @@
  */
 
 #include "apphook.h"
-#include "timeutils.h"
-#include "timeutils.h"
+#include "timeutils/timeutils.h"
+#include "timeutils/timeutils.h"
 #include "logstamp.h"
 #include <criterion/criterion.h>
 #include <stdlib.h>
